CountriesData class combined with Country class
